Add files via upload
this is an example of what this program can create. If you don't have a monstrous computer you will need to wait for a minute to open the file. This file was simplified with meshmaker and turned into a 3mf for your ease of use. The file is just one forth of the whole 3d structure.