I fixed some issues, more in description
I changed the Threshold and pixel weight values because pixel weight was too low, making only the insides of the Mandelbrot set visible. Threshold was causing a similar issue. I have (even though it causes a less efficient program) changed the loops so that the z symmetry is also calculated.